REQ[#68966]: 磁盘清理接口修改
[Description]: 1. 磁盘清理接口增加操作日志
[Root Cause]:
[Solution]:
[Test Suggestion]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/config/controller/SysDiskCleanupController.java
@@ -4,6 +4,9 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.core.metadata.OrderItem;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.yjh.platform.common.Constant;
+import com.yjh.platform.common.logs.Logs;
+import com.yjh.platform.common.logs.LogsRecord;
 import com.yjh.platform.common.result.BusinessException;
 import com.yjh.platform.common.result.Result;
 import com.yjh.platform.common.result.ResultCodeEnum;
@@ -14,6 +17,8 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.MapUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.math.NumberUtils;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.bind.annotation.*;
 
@@ -38,13 +43,16 @@ public class SysDiskCleanupController {
 
     private final ISysDiskCleanupService sysDiskCleanupService;
     private final RedisTemplate redisTemplate;
+    private final LogsRecord logsRecord;
 
-    public SysDiskCleanupController(ISysDiskCleanupService sysDiskCleanupService, RedisTemplate redisTemplate) {
+    public SysDiskCleanupController(ISysDiskCleanupService sysDiskCleanupService, RedisTemplate redisTemplate, LogsRecord logsRecord) {
         this.sysDiskCleanupService = sysDiskCleanupService;
         this.redisTemplate = redisTemplate;
+        this.logsRecord = logsRecord;
     }
 
     @ApiOperation(value = "查询历史磁盘清理任务")
+    @Logs(title = "磁盘清理记录查询", content = "查询历史磁盘清理任务", logType = 1, authority = "1234")
     @GetMapping(value = "/selectByPage")
     public Result selectByPage(@RequestParam(value = "pageNum", required = false, defaultValue = "1") int pageNum,
         @RequestParam(value = "pageSize", required = false, defaultValue = "20") int pageSize) {
@@ -70,6 +78,7 @@ public class SysDiskCleanupController {
     }
 
     @ApiOperation(value = "查询正在执行的磁盘清理任务")
+    @Logs(title = "磁盘清理", content = "查询正在执行的磁盘清理任务", logType = 1, authority = "1234")
     @GetMapping(value = "/runningCleanup")
     public Result runningCleanup() {
         Result result = new Result();
@@ -85,11 +94,17 @@ public class SysDiskCleanupController {
     }
 
     @ApiOperation(value = "新增磁盘清理任务")
+    @Logs(title = "磁盘清理新增", content = "新增磁盘清理任务", logType = 2, authority = "1234")
     @PostMapping(value = "/add")
     public Result add(@RequestBody SysDiskCleanup sysDiskCleanup, HttpServletRequest request) {
         Result result = new Result();
         try {
             String userId = request.getHeader("userId");
+            if (Constant.MANAGER_USER_ID != NumberUtils.toLong(userId, Constant.MANAGER_USER_ID)) {
+                result.setMessage(ResultCodeEnum.CODE10008.getCode(), "仅系统默认管理员可进行磁盘清理操作");
+                return result;
+            }
+
             String userName = (String)redisTemplate.opsForHash().get("userInfo:" + userId, "userName");
             sysDiskCleanup.setCreator(userName);
             result.setData(sysDiskCleanupService.addTask(sysDiskCleanup));
@@ -103,13 +118,19 @@ public class SysDiskCleanupController {
     }
 
     @ApiOperation(value = "磁盘清理任务执行完成确认")
+    @Logs(title = "磁盘清理确认", content = "磁盘清理任务执行完成确认", logType = 5, authority = "1234")
     @PostMapping(value = "/confirm")
-    public Result confirm(@RequestBody Map<String, Object> cleanMap) {
+    public Result confirm(@RequestBody Map<String, Object> cleanMap, HttpServletRequest request) {
         Result result = new Result();
         try {
+            if (Constant.MANAGER_USER_ID != NumberUtils.toLong(request.getHeader("userId"), Constant.MANAGER_USER_ID)) {
+                result.setMessage(ResultCodeEnum.CODE10008.getCode(), "仅系统默认管理员可进行磁盘清理确认");
+                return result;
+            }
+
             int cleanId = MapUtils.getIntValue(cleanMap, "id");
             SysDiskCleanup cleanup = sysDiskCleanupService.getById(cleanId);
-            if (cleanup.getCleanStatus() != 0) {
+            if (cleanup == null || cleanup.getCleanStatus() != 0) {
                 result.setCode(ResultCodeEnum.CODE10009.getCode(), "当前任务未完成或已经确认，无法进行确认");
                 return result;
             }
@@ -129,12 +150,23 @@ public class SysDiskCleanupController {
 
     @ApiOperation(value = "恢复数据")
     @PostMapping(value = "/recovery")
-    public Result recovery(@RequestBody Map<String, Object> cleanMap, @RequestParam(value = "id") int cleanId,
-        @RequestParam(value = "type") int type) {
+    public Result recovery(@RequestBody Map<String, Object> cleanMap, HttpServletRequest request) {
         Result result = new Result();
         try {
-            result.setData(
-                sysDiskCleanupService.recoveryTask(MapUtils.getIntValue(cleanMap, "id"), MapUtils.getIntValue(cleanMap, "type")));
+            String userId = request.getHeader("userId");
+            String userName = (String)redisTemplate.opsForHash().get("userInfo:" + userId, "userName");
+            int type = MapUtils.getIntValue(cleanMap, "type");
+            String name = type == 1 ? "数据库记录" : "数据文件";
+
+            if (Constant.MANAGER_USER_ID != NumberUtils.toLong(userId, Constant.MANAGER_USER_ID)) {
+                result.setMessage(ResultCodeEnum.CODE10008.getCode(), "仅系统默认管理员可进行数据恢复");
+                logsRecord.LoginLogsSend(request, "3", "磁盘清理数据恢复", "恢复磁盘清理删除的" + name, userName, userId, 2);
+                return result;
+            }
+
+            logsRecord.LoginLogsSend(request, "3", "磁盘清理数据恢复", "恢复磁盘清理删除的" + name, userName, userId, 1);
+
+            result.setData(sysDiskCleanupService.recoveryTask(MapUtils.getIntValue(cleanMap, "id"), type));
         } catch (BusinessException b) {
             result.setCode(b.getCode(), b.getMessage());
         } catch (Exception e) {
@@ -145,10 +177,16 @@ public class SysDiskCleanupController {
     }
 
     @ApiOperation(value = "删除备份")
+    @Logs(title = "磁盘清理备份删除", content = "删除磁盘清理备份的数据文件", logType = 4, authority = "1234")
     @PostMapping(value = "/deleteBack")
-    public Result deleteBack(@RequestBody Map<String, Object> cleanMap) {
+    public Result deleteBack(@RequestBody Map<String, Object> cleanMap, HttpServletRequest request) {
         Result result = new Result();
         try {
+            if (Constant.MANAGER_USER_ID != NumberUtils.toLong(request.getHeader("userId"), Constant.MANAGER_USER_ID)) {
+                result.setMessage(ResultCodeEnum.CODE10008.getCode(), "仅系统默认管理员可删除备份");
+                return result;
+            }
+
             result.setData(sysDiskCleanupService.deleteBack(MapUtils.getIntValue(cleanMap, "id")));
         } catch (BusinessException b) {
             result.setCode(b.getCode(), b.getMessage());
